backend/requirements.txt
fastapi
uvicorn
python-multipart
semgrep
fpdf
google-generativeai